fix float -> double

+++ QLNH_Winform/DAO/FoodDAO.cs
@@ -94,14 +94,14 @@ namespace QLNH_Winform.DAO
             return list;
         }
 
-        public bool InsertFood(string name, int id, float price)
+        public bool InsertFood(string name, int id, double price)
         {
             string query = string.Format("INSERT Food ( name, idCategory, price) VALUES (N'{0}', {1}, {2})", name, id, price);
             int result = DataProvider.Instance.ExecuteNonQuery(query);
 
             return result > 0;
         }
-        public bool UpdateFood(int idFood, string name, int id, float price)
+        public bool UpdateFood(int idFood, string name, int id, double price)
         {
             string query = string.Format("UPDATE Food SET name = N'{0}', idCategory = {1}, price = {2} WHERE id = {3}", name, id, price, idFood);
             int result = DataProvider.Instance.ExecuteNonQuery(query);
