+++ QLNH_Winform/Forms/FormGoiMon.cs
@@ -67,7 +67,7 @@ namespace QLNH_Winform.Forms
         {
             lsvBill.Items.Clear();
             List<MenuFood> listBillInfo = MenuFoodDAO.Instance.GetListMenuByTable(id);
-            float totalPrice = 0;
+            double totalPrice = 0;
             foreach (MenuFood item in listBillInfo)
             {
                 ListViewItem lsvitem = new ListViewItem(item.FoodName.ToString());
@@ -151,7 +151,7 @@ namespace QLNH_Winform.Forms
                 DialogResult res = MessageBox.Show(string.Format("Bạn có chắc thanh toán hóa đơn cho bàn {0} ?\nHóa đơn đã được giảm {2}%\nTổng tiền cần thanh toán là: {1} VND", table.Name, finalTotalPrice, discount), "Thông báo", MessageBoxButtons.OKCancel, MessageBoxIcon.Question);
                 if (res == DialogResult.OK)
                 {
-                    BillDAO.Instance.CheckOut(idBill, discount, (float)finalTotalPrice);
+                    BillDAO.Instance.CheckOut(idBill, discount, (double)finalTotalPrice);
                     ShowBill(table.ID);
                     LoadTable();
                 }
